llm/requirements.txt
requests
python-dotenv
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
